+++ core/main.ts
@@ -37,6 +37,7 @@ export * from "./src/ecs/systems/ui/RenderSystem";
 export * from "./src/ecs/entities/buttons/ConnectButton";
 export * from "./src/ecs/entities/buttons/FullscreenButton";
 export * from "./src/ecs/entities/buttons/HomeButton";
+export * from "./src/ecs/entities/buttons/ShopButton";
 export * from "./src/ecs/entities/characters/Skelly";
 export * from "./src/ecs/entities/characters/Spaceship";
 export * from "./src/ecs/entities/characters/Zombie";
@@ -58,7 +59,7 @@ export * from "./src/ecs/entities/ui/Joystick";
 export * from "./src/ecs/renderables/Button";
 export * from "./src/ecs/renderables/DebugBounds";
 export * from "./src/ecs/renderables/HealthBar";
-export * from "./src/ecs/entities/ui/BuyMenu";
+export * from "./src/ecs/entities/ui/Shop";
 export * from "./src/ecs/entities/ui/Minimap";
 export * from "./src/ecs/entities/ui/Scoreboard";
 export * from "./src/ecs/entities/ui/ScorePanel";
