+++ docs/index.css
@@ -13,6 +13,7 @@ html {
 }
 
 body, html {
+  touch-action: none;
   margin: 0;
   padding: 0;
   overflow: hidden;
