+++ games/home/Home.ts
@@ -1,6 +1,6 @@
 import {
   Background, ZombieSpawnSystem, FloorTiles, GunSystem, InviteStone,
-  IsometricGame, LineWall, Portal, SkellySpawnSystem, BuyMenu
+  IsometricGame, LineWall, Portal, SkellySpawnSystem, Shop
 } from "@piggo-gg/core";
 
 export const Home = IsometricGame({
@@ -11,7 +11,7 @@ export const Home = IsometricGame({
     entities: [
       Background({ img: "stars.png" }),
 
-      BuyMenu(),
+      Shop(),
 
       Portal({ pos: { x: 544, y: 144 }, game: "aram", tint: 0xff0055 }),
       Portal({ pos: { x: -480, y: 144 }, game: "legends", tint: 0x00ffcc }),
