+++ games/strike/Strike.ts
@@ -1,5 +1,5 @@
 import {
-  Background, BuyMenu, FloorCollidersArray, FloorTilesArray, GunSystem,
+  Background, Shop, FloorCollidersArray, FloorTilesArray, GunSystem,
   HealthBarSystem, HomeButton, IsometricGame, Minimap,
   ScorePanel, Scoreboard, SightSystem, isMobile
 } from "@piggo-gg/core";
@@ -16,7 +16,7 @@ export const Strike = IsometricGame({
       Background(),
       ScorePanel(),
       Scoreboard(),
-      BuyMenu(),
+      Shop(),
       FloorTilesArray(80, StrikeTileMap),
       ...FloorCollidersArray(80, StrikeTileMap),
       ... (isMobile() ? [] : [Minimap(80, StrikeTileMap)])
